Move socket creation for Firefox compatibility
If the websocket opens before the open event listener is assigned, the open
event is never fired.

The reason there was enough time for the web socket to connect seems to be the
prompt we do for the username. It takes time and probably interrupts the main
thread, allowing the connection task to be completed.

This seems to be an implementation difference between Firefox and Chrome. It
could be that Chrome will fire the event when it is added, even if the
connection is already made. It also might be that Chrome does not try to create
the connection while the prompt is open.